+++ src/ProcessMaker/Nayra/Model/Activity.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace ProcessMaker\Nayra\Model;
+
+use ProcessMaker\Models\ActivityActivatedEvent;
+use ProcessMaker\Nayra\Bpmn\ActivityTrait;
+use ProcessMaker\Nayra\Bpmn\FlowNodeTrait;
+use ProcessMaker\Nayra\Bpmn\ProcessTrait;
+use ProcessMaker\Nayra\Contracts\Bpmn\ActivityInterface;
+use ProcessMaker\Nayra\Contracts\Bpmn\FlowCollectionInterface;
+use ProcessMaker\Nayra\Contracts\Bpmn\FlowInterface;
+use ProcessMaker\Nayra\Contracts\Bpmn\FlowNodeInterface;
+use ProcessMaker\Nayra\Contracts\Bpmn\GatewayInterface;
+use ProcessMaker\Nayra\Contracts\Bpmn\ProcessInterface;
+use ProcessMaker\Nayra\Contracts\Repositories\FlowRepositoryInterface;
+
+/**
+ * Activity implementation.
+ *
+ * @package ProcessMaker\Models
+ */
+class Activity implements ActivityInterface
+{
+    use ActivityTrait;
+
+    /**
+     * Array map of custom event classes for the bpmn element.
+     *
+     * @return array
+     */
+    protected function getBpmnEventClasses()
+    {
+        return [
+            ActivityInterface::EVENT_ACTIVITY_ACTIVATED => ActivityActivatedEvent::class,
+        ];
+    }
+}
