+++ tests/ProcessMaker/Models/LocalFlowNodeTrait.php
@@ -2,6 +2,7 @@
 
 namespace ProcessMaker\Models;
 
+use ProcessMaker\Nayra\Bpmn\FlowNodeTrait;
 use ProcessMaker\Nayra\Contracts\Bpmn\FlowInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\FlowNodeInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\GatewayInterface;
@@ -15,43 +16,6 @@ use ProcessMaker\Nayra\Contracts\Repositories\FlowRepositoryInterface;
 trait LocalFlowNodeTrait
 {
 
-    /**
-     * @param FlowNodeInterface $target
-     * @param FlowRepositoryInterface $flowRepository
-     * @param array $properties
-     * @return $this
-     */
-    public function createFlowTo(FlowNodeInterface $target, FlowRepositoryInterface $flowRepository, $properties=[])
-    {
-        $flow = $flowRepository->createFlowInstance();
-        $flow->setSource($this);
-        $flow->setTarget($target);
-        $flow->setProperties($properties);
-        $this->addProperty(FlowNodeInterface::BPMN_PROPERTY_OUTGOING, $flow);
-        $target->addProperty(FlowNodeInterface::BPMN_PROPERTY_INCOMING, $flow);
-        if (!empty($properties[FlowInterface::BPMN_PROPERTY_IS_DEFAULT])) {
-            $this->setProperty(GatewayInterface::BPMN_PROPERTY_DEFAULT, $flow);
-        }
-        return $this;
-    }
+    use FlowNodeTrait;
 
-    /**
-     * Get flows.
-     *
-     * @return array
-     */
-    public function getFlows()
-    {
-        return $this->getProperty(FlowNodeInterface::BPMN_PROPERTY_OUTGOING);
-    }
-
-    /**
-     * Get the incoming flows.
-     *
-     * @return FlowCollectionInterface
-     */
-    public function getIncomingFlows()
-    {
-        return $this->getProperty(FlowNodeInterface::BPMN_PROPERTY_INCOMING);
-    }
 }
